requirements.txt
appnope==0.1.0
cling==2.3.1
cookies==2.2.1
coverage==4.0.3
coveralls==1.1
decorator==4.0.9
dj-database-url==0.4.0
dj-static==0.0.6
Django==1.9.2
django-allauth==0.24.1
django-appconf==1.0.1
django-bootstrap3==6.2.2
django-facebook==6.0.3
django-imagekit==3.3
django-social-share==0.3.0
docopt==0.6.2
factory-boy==2.6.1
fake-factory==0.5.3
funcsigs==0.4
gnureadline==6.3.3
gunicorn==19.4.5
httplib2==0.9.2
ipdb==0.8.1
ipython==4.1.1
ipython-genutils==0.1.0
mock==1.3.0
oauth2==1.9.0.post1
oauthlib==1.0.3
path.py==8.1.2
pbr==1.8.1
pexpect==4.0.1
pickleshare==0.6
pilkit==1.1.13
Pillow==3.1.1
psycopg2==2.6.1
ptyprocess==0.5.1
python-openid==2.2.5
requests==2.9.1
requests-oauthlib==0.6.0
responses==0.5.1
selenium==2.52.0
simplegeneric==0.8.1
six==1.10.0
static3==0.6.1
traitlets==4.1.0
Unidecode==0.4.19
wheel==0.24.0
